+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/presentation/dto/response/MemberResponseDto.java
@@ -1,4 +1,4 @@
-package com.nextdoor.nextdoor.domain.member.controller.dto.response;
+package com.nextdoor.nextdoor.domain.member.presentation.dto.response;
 
 import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import com.nextdoor.nextdoor.domain.member.enums.Gender;
